Restructure project into focused single-responsibility modules
- Split monolithic dboperations.py (549 lines) into:
  - database.py: Database query class only
  - builder.py: DB creation and YAML parsing
  - downloader.py: network download logic
- Split monolithic material.py (526 lines) into:
  - material.py: Material class only
  - optical_data.py: RefractiveIndexData, FormulaRefractiveIndexData,
    TabulatedRefractiveIndexData, ExtinctionCoefficientData
  - exceptions.py: FormulaNotImplemented, NoExtinctionCoefficient
  - models.py: Shelf, Book, Page, Entry namedtuples
  - _constants.py: RII_DATABASE_URL
- Add pyproject.toml for pip-installable packaging
- Update __init__.py to expose Database and Material as public API
- Move Tutorial.ipynb -> examples/, ER.PNG -> docs/
- Move .gitignore from package folder to repo root
- Add tests/ with unit tests for optical_data, material, and database
