+++ refractivesqlite/__init__.py
@@ -1,1 +1,4 @@
+from refractivesqlite.database import Database
+from refractivesqlite.material import Material
 
+__all__ = ["Database", "Material"]
