+++ refractivesqlite/_constants.py
@@ -1,0 +1,4 @@
+RII_DATABASE_URL = (
+    "https://refractiveindex.info/download/database/"
+    "rii-database-2019-02-11.zip"
+)
