+++ refractivesqlite/exceptions.py
@@ -1,0 +1,14 @@
+class FormulaNotImplemented(Exception):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)
+
+
+class NoExtinctionCoefficient(Exception):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)
